+++ geojson_pydantic/__init__.py
@@ -1,3 +1,14 @@
 """geojson-pydantic."""
 
-version = "0.2.3"
+from .features import Feature, FeatureCollection  # noqa
+from .geometries import (  # noqa
+    GeometryCollection,
+    LineString,
+    MultiLineString,
+    MultiPoint,
+    MultiPolygon,
+    Point,
+    Polygon,
+)
+
+version = "0.3.0"