+++ geojson_pydantic/geometries.py
@@ -1,12 +1,12 @@
 """pydantic models for GeoJSON Geometry objects."""
 
 import abc
-from typing import Any, List, Tuple, Union
+from typing import Any, List, Union
 
 from pydantic import BaseModel, Field, ValidationError, validator
 from pydantic.error_wrappers import ErrorWrapper
 
-from .utils import NumType
+from geojson_pydantic.types import Position
 
 
 class _GeometryBase(BaseModel, abc.ABC):
@@ -19,60 +19,58 @@ class _GeometryBase(BaseModel, abc.ABC):
         return self.dict()
 
 
-Coordinate = Union[Tuple[NumType, NumType], Tuple[NumType, NumType, NumType]]
-
-Position = Coordinate
-
-
 class Point(_GeometryBase):
     """Point Model"""
 
     type: str = Field("Point", const=True)
-    coordinates: Coordinate
+    coordinates: Position
 
 
 class MultiPoint(_GeometryBase):
     """MultiPoint Model"""
 
     type: str = Field("MultiPoint", const=True)
-    coordinates: List[Coordinate]
+    coordinates: List[Position]
 
 
 class LineString(_GeometryBase):
     """LineString Model"""
 
     type: str = Field("LineString", const=True)
-    coordinates: List[Coordinate] = Field(..., min_items=2)
+    coordinates: List[Position] = Field(..., min_items=2)
 
 
 class MultiLineString(_GeometryBase):
     """MultiLineString Model"""
 
     type: str = Field("MultiLineString", const=True)
-    coordinates: List[List[Coordinate]]
+    coordinates: List[List[Position]]
 
 
 class Polygon(_GeometryBase):
     """Polygon Model"""
 
     type: str = Field("Polygon", const=True)
-    coordinates: List[List[Coordinate]]
+    coordinates: List[List[Position]]
 
     @validator("coordinates")
-    def check_coordinates(cls, coords):
+    def check_coordinates(cls, polygon):
         """Validate that Polygon coordinates pass the GeoJSON spec"""
-        if any([len(c) < 4 for c in coords]):
+        if len(polygon) == 0:
+            raise ValueError("Polygon must have a minimum one ring")
+        if any([len(ring) < 4 for ring in polygon]):
             raise ValueError("All linear rings must have four or more coordinates")
-        if any([c[-1] != c[0] for c in coords]):
+        if any([ring[-1] != ring[0] for ring in polygon]):
             raise ValueError("All linear rings have the same start and end coordinates")
-        return coords
+
+        return polygon
 
 
 class MultiPolygon(_GeometryBase):
     """MultiPolygon Model"""
 
     type: str = Field("MultiPolygon", const=True)
-    coordinates: List[List[List[Coordinate]]]
+    coordinates: List[List[List[Position]]]
 
 
 Geometry = Union[Point, MultiPoint, LineString, MultiLineString, Polygon, MultiPolygon]